Support single core Zynq devices (MiniZed).
Summary: On Zynq platforms, get the CPU count from the SCU_CONFIG register.  This is necessary to boot on Xilinx' new single core chips (such as on the MiniZed).  I don't have access to MiniZed yet but I have been debugging on Xilinx' port of qemu which acts like a single core Zynq so I have been using this patch for a while.

Test Plan: Tested on Zedboard and qemu.

Reviewers: jmg

Subscribers: andrew

Differential Revision: https://reviews.freebsd.org/D14429